move 'AppointmentsPage' directory to 'pages' directory
