+++ src/applications/vaos/appointment-list/pages/AppointmentsPage/index.jsx
@@ -11,19 +11,23 @@ import WarningNotification from '../../../components/WarningNotification';
 import { selectFeatureBreadcrumbUrlUpdate } from '../../../redux/selectors';
 import { APPOINTMENT_STATUS } from '../../../utils/constants';
 import { scrollAndFocus } from '../../../utils/scrollAndFocus';
-import RequestedAppointmentsPage from '../../pages/RequestedAppointmentsPage/RequestedAppointmentsPage';
+import RequestedAppointmentsPage from '../RequestedAppointmentsPage/RequestedAppointmentsPage';
 import { selectPendingAppointments } from '../../redux/selectors';
-import AppointmentListNavigation from '../AppointmentListNavigation';
-import PageLayout from '../PageLayout';
-import PastAppointmentsList from '../PastAppointmentsList';
-import ScheduleNewAppointment from '../ScheduleNewAppointment';
+// import AppointmentListNavigation from '../AppointmentListNavigation';
+// import PageLayout from '../PageLayout';
+// import PastAppointmentsList from '../PastAppointmentsList';
+// import ScheduleNewAppointment from '../ScheduleNewAppointment';
 // import CernerTransitionAlert from '../../../components/CernerTransitionAlert';
 // import { selectPatientFacilities } from '~/platform/user/cerner-dsot/selectors';
 import ReferralTaskCardWithReferral from '../../../referral-appointments/components/ReferralTaskCardWithReferral';
 import { routeToCCPage } from '../../../referral-appointments/flow';
 import { useIsInCCPilot } from '../../../referral-appointments/hooks/useIsInCCPilot';
 import { setFormCurrentPage } from '../../../referral-appointments/redux/actions';
-import UpcomingAppointmentsPage from '../../pages/UpcomingAppointmentsPage/UpcomingAppointmentsPage';
+import UpcomingAppointmentsPage from '../UpcomingAppointmentsPage/UpcomingAppointmentsPage';
+import PageLayout from '../../components/PageLayout';
+import ScheduleNewAppointment from '../../components/ScheduleNewAppointment';
+import AppointmentListNavigation from '../../components/AppointmentListNavigation';
+import PastAppointmentsListNew from '../../components/PastAppointmentsList';
 
 function renderWarningNotification() {
   return (props, childContent) => {
@@ -192,7 +196,7 @@ export default function AppointmentsPage() {
           <RequestedAppointmentsPage hasTypeChanged={hasTypeChanged} />
         </Route>
         <Route path="/past">
-          <PastAppointmentsList hasTypeChanged={hasTypeChanged} />
+          <PastAppointmentsListNew hasTypeChanged={hasTypeChanged} />
         </Route>
       </Switch>
     </PageLayout>
