+++ src/applications/vaos/referral-appointments/components/RequestsList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import RequestAppointmentLayout from '../../appointment-list/components/AppointmentsPage/RequestAppointmentLayout';
+import RequestAppointmentLayout from '../../components/RequestAppointmentLayout';
 import { APPOINTMENT_STATUS } from '../../utils/constants';
 import InfoAlert from '../../components/InfoAlert';
 import ScheduleAppointmentLink from '../../appointment-list/components/ScheduleAppointmentLink';
